requirements.txt
boto3
python-dateutil
pytimeparse
fiona
Pillow
pyproj
rasterio
shapely
tqdm